# Optimize Parquet reading by reading batches rather than creating a copy