--- a/src/Parquet.chpl
+++ b/src/Parquet.chpl
@@ -9,7 +9,8 @@ module Parquet {
   }
 
   private config const ROWGROUPS = 512*1024*1024 / numBytes(int); // 512 mb of int64
-  private config const batchSize = getEnvInt("ARKOUDA_PARQUET_BATCH_SIZE", 8192);
+  // Undocumented for now, just for internal experiments
+  private config const batchSize = getEnvInt("ARKOUDA_SERVER_PARQUET_BATCH_SIZE", 8192);
 
   extern var ARROWINT64: c_int;
   extern var ARROWINT32: c_int;