Today, to read in a Parquet file, we are reading the Parquet file into a temporary array and then copying the contents of the temporary array into the Chapel array.

This optimization reads "batches" of 8,192 elements of int64 directly into the Chapel array, avoiding the extra array copy and need to copy elements one by one. For int32 Parquet columns, batches of single elements are read into a temporary variable and then copied to the Chapel array to avoid having to keep a temporary copy of the column around.

Here are reading performance numbers collected on chapcs
Parquet type | Current | This Branch
-- | -- | --
int64 | 0.69 GiB/s | 1.41 GiB/s
int32 | 0.65 GiB/s | 1.29 GiB/s

So, we see about a 2x improvement in performance case and also remove the extra array copy, so have a lower memory load.

Additionally, this test adds an int32 Parquet file under `test/resources` to test int32 reading in the Chapel unit testing.